Add support for remembering location permission requests
Every time users visit a page which asks for location permissions a
prompt dialog is shown. This is suboptimal as it forces users to
accept/decline the offer every single time.

Most browsers out there include a checkbox in the permission dialog
offering users to remember their decision. We're basically adding
that functionality.

We're only storing location permissions. We considered adding
desktop notifications (as it should be straightforward from now on)
but we aren't doing it because we don't really support them (we don't
show desktop notifications at the moment).